+++ .nvmrc
@@ -1,0 +1,1 @@
+7.7.1
