+++ package.json
@@ -71,6 +71,7 @@
     "history": "^4.5.1",
     "normalize.css": "^5.0.0",
     "react": "^15.3.1",
+    "react-contenteditable": "^2.0.2",
     "react-dnd": "^2.1.4",
     "react-dnd-html5-backend": "^2.1.2",
     "react-dom": "^15.3.1",
