+++ src/components/baseline/index.js
@@ -5,3 +5,4 @@
 export { Button } from './components/Button'
 export { Card } from './components/Card'
 export { Icon } from './components/Icon';
+export { TextEditor } from './components/TextEditor';
