+++ src/routes/project/components/Headline.js
@@ -1,13 +1,15 @@
 import React from 'react';
+import { TextEditor } from 'components/baseline';
+
 
 const Headline = ({name, client}) => {
-  return(
+  console.log(TextEditor);
+  return (
     <div className='Headline'>
       <div className="content">
-        <h1>{name}</h1>
-        <h3>{client ? client : 'Client Name'}</h3> 
-
-        <p className='date'>November 20, 2016</p>
+        <TextEditor html={`<h1>${name}</h1>`} />
+        <TextEditor html={`<h3>${client ? client : 'Client Name'}</h3>`} />
+        <TextEditor html={`<p className=\"date\">November 20, 2016</p>`} />
       </div>
     </div>
   )
