+++ src/utils/index.js
@@ -67,3 +67,5 @@ export function parseQueryString(str) {
 
       return ret;
     }
+  
+  export const stripTags = (str) => str.replace(/(<([^>]+)>)/ig,"");
